+++ versione3/Package/globals.py
@@ -1,0 +1,26 @@
+# Variabili globali
+
+
+game_iteration = 8000
+cpu_02 = 0.7
+
+# Coordinate Luke
+x_global = 5
+y_global = 20
+side_global = 93
+
+"""
+# Coordinate Chrii
+x_global = 6
+y_global = 23
+side_global = 94
+"""
+
+chri_webhook = "https://discord.com/api/webhooks/1184788603470090320/jqkYHRC-y7P920AYLdB1e08g1WPLWzIcelEssk1tG23VXXE2kvgsNUQUhg5q7fYZ86hU"
+luke_webhook = "https://discord.com/api/webhooks/1184915602431823944/3HyIjU1u40NnJUVpzzfBAEOAHZ9EZkiUwEeViffQLOxUwLpU25-dRR0-LJlx2snXbdsH"
+num_righe = 6
+num_colonne = 5
+
+
+
+
